refactor(client): reuse account wallet binding
