+++ packages/client/src/decorators/index.ts
@@ -9,11 +9,7 @@ import {
   type PublicAccountActions,
   type SecureAccountActions,
 } from './account';
-import {
-  type AnalyticsActions,
-  analyticsActions,
-  type SecureAnalyticsActions,
-} from './analytics';
+import { type AnalyticsActions, analyticsActions } from './analytics';
 import { type DataActions, dataActions } from './data';
 import { type DiscoveryActions, discoveryActions } from './discovery';
 import {
@@ -49,7 +45,7 @@ export type PublicActions = Prettify<
 export type SecureActions = Prettify<
   DiscoveryActions &
     DataActions &
-    SecureAnalyticsActions &
+    AnalyticsActions &
     SecurePerpsActions &
     SecureAccountActions &
     SecureRewardsActions &
